+++ betahaus/viewcomponent/models.py
@@ -7,8 +7,12 @@ from pyramid.traversal import find_interface
 from betahaus.viewcomponent.interfaces import IViewAction
 from betahaus.viewcomponent.interfaces import IViewGroup
 from betahaus.viewcomponent.compat import string_types
+from pyramid.threadlocal import get_current_registry
 
 logger = logging.getLogger(__name__)
+
+_marker = object()
+_empty_vals = ('', None)
 
 
 @implementer(IViewGroup)
@@ -25,9 +29,15 @@ class ViewGroup(object):
         self._order = []
         self._data = {}
     
-    def __call__(self, context, request, **kw):
-        va_output = [x(context, request, **kw) for x in self.values()]
-        return "".join([x for x in va_output if x])
+    def __call__(self, context, request,
+                 as_type = None, spacer = "", empty_val = _marker, **kw):
+        if as_type is None:
+            return spacer.join([x for x in self.as_generator(context, request, empty_val = empty_val, **kw)])
+        try:
+            type_method = getattr(self, 'as_%s' % as_type)
+        except AttributeError:
+            raise ValueError("as_type must be a valid output type or None, was: %r" % as_type)
+        return type_method(context, request, empty_val = empty_val, **kw)
 
     def __getitem__(self, key):
         return self._data[key]
@@ -36,6 +46,8 @@ class ViewGroup(object):
         assert isinstance(value, ViewAction)
         value.parent = self
         if value.priority is not None:
+            if key in self.order:
+                self.order.remove(key)
             index = len(self)
             for va in reversed(self.values()):
                 if va.priority is not None and va.priority <= value.priority:
@@ -43,7 +55,8 @@ class ViewGroup(object):
                 index -= 1
             self.order.insert(index, key)
         else:
-            self.order.append(key)
+            if key not in self.order:
+                self.order.append(key)
         self._data[key] = value
 
     def __delitem__(self, key):
@@ -81,10 +94,6 @@ class ViewGroup(object):
             new_order.extend(handle_keys)
         self._order = new_order
 
-    def get_context_vas(self, context, request): # pragma: no cover
-        # backward compat, not needed
-        return self.values()
-
     def add(self, view_action):
         self[view_action.name] = view_action
 
@@ -99,6 +108,29 @@ class ViewGroup(object):
 
     def items(self):
         return [(name, self[name]) for name in self.order]
+
+    def as_generator(self, context, request, empty_val = _marker, **kw):
+        for x in self.values():
+            res = x(context, request, **kw)
+            if res in _empty_vals:
+                if empty_val != _marker:
+                    yield empty_val
+            else:
+                yield res
+
+    def as_dict(self, context, request, empty_val = _marker, **kw):
+        va_output = {}
+        for (k, v) in self.items():
+            res = v(context, request, **kw)
+            if res in _empty_vals:
+                if empty_val != _marker:
+                    va_output[k] = empty_val
+            else:
+                va_output[k] = res
+        return va_output
+
+    def as_list(self, context, request, empty_val = _marker, **kw):
+        return list(self.as_generator(context, request, empty_val = empty_val, **kw))
 
     def __repr__(self): # pragma : no cover
         klass = self.__class__
@@ -136,3 +168,18 @@ class ViewAction(object):
         klass = self.__class__
         classname = '%s.%s' % (klass.__module__, klass.__name__)
         return "<%s '%s'>" % (classname, self.name)
+
+
+def add_view_action(_callable, group_name, name, priority = None, registry = None, **kwargs):
+    """ Create a new view action and possibly add a view group if it doesn't exist.
+        Use the decorator or the directive instead of this method directly.
+    """
+    if registry is None: # pragma : no cover
+        registry = get_current_registry()
+    view_group = registry.queryUtility(IViewGroup, name = group_name)
+    if view_group is None:
+        view_group = ViewGroup(group_name)
+        registry.registerUtility(view_group, IViewGroup, name = group_name)
+    va = ViewAction(_callable, name, priority = priority, **kwargs)
+    view_group.add(va)
+    return va